feat: 동영상 플레이어 단축키 추가
- 재생 / 일시정지
- 음소거 / 음소거 해제
- 전체화면 / 전체화면 해제
