+++ src/components/video/videoPlayer.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useRef, useState } from 'react';
+import { KeyboardEvent, useRef, useState } from 'react';
 import { debounce } from '@/utils/debounce';
 import Spinner from './spinner';
 import { videoPlayerStyle } from '@/styles/video.css';
@@ -19,6 +19,7 @@ export default function VideoPlayer({
 }: VideoPlayerProps) {
   const containerRef = useRef<HTMLElement>(null);
   const videoRef = useRef<HTMLVideoElement>(null);
+  const playPromiseRef = useRef<Promise<void>>(null);
 
   const [isPanelShown, setIsPanelShown] = useState<boolean>(false);
   const [isPlaying, setIsPlaying] = useState<boolean>(false);
@@ -26,6 +27,8 @@ export default function VideoPlayer({
   const [duration, setDuration] = useState<number>(0);
   const [bufferedProgress, setBufferedProgress] = useState<number>(0);
   const [isBuffering, setIsBuffering] = useState<boolean>(false);
+  const [isMuted, setIsMuted] = useState<boolean>(false);
+  const [isFullscreen, setIsFullscreen] = useState<boolean>(false);
 
   const debouncedHidePanelRef = useRef(debounce(() => setIsPanelShown(false), 3000));
 
@@ -54,12 +57,107 @@ export default function VideoPlayer({
     }
   };
 
+  const playVideo = () => {
+    if (!videoRef.current) {
+      return;
+    }
+
+    if (playPromiseRef.current) {
+      return;
+    }
+
+    const playPromise = videoRef.current.play();
+
+    if (playPromise !== undefined) {
+      playPromiseRef.current = playPromise;
+
+      playPromise.then(() => {
+        playPromiseRef.current = null;
+      }).catch(() => {
+        setIsPlaying(false);
+        playPromiseRef.current = null;
+      });
+    }
+  };
+
+  const pauseVideo = () => {
+    if (!videoRef.current) {
+      return;
+    }
+
+    videoRef.current.pause();
+  };
+
+  const handlePlayPause = () => {
+    if (isPlaying) {
+      pauseVideo();
+    }
+    else {
+      playVideo();
+    }
+  };
+
+  const handleMuteVolume = () => {
+    if (!videoRef.current) {
+      return;
+    }
+
+    if (videoRef.current.muted) {
+      videoRef.current.muted = false;
+      setIsMuted(false);
+    }
+    else {
+      videoRef.current.muted = true;
+      setIsMuted(true);
+    }
+  };
+
+  const handleFullscreen = () => {
+    if (!containerRef.current) {
+      return;
+    }
+
+    if (!document.fullscreenElement) {
+      containerRef.current.requestFullscreen();
+      setIsFullscreen(true);
+    }
+    else {
+      document.exitFullscreen();
+      setIsFullscreen(false);
+    }
+  };
+
+  const handleShortcut = (e: KeyboardEvent<HTMLElement>) => {
+    switch (e.key.toLowerCase()) {
+      case ' ': {
+        handlePlayPause();
+        break;
+      }
+
+      case 'm': {
+        handleMuteVolume();
+        break;
+      }
+
+      case 'enter': {
+        handleFullscreen();
+        break;
+      }
+
+      default: {
+        break;
+      }
+    }
+  };
+
   return (
     <figure
       className={`${videoPlayerStyle.container}${isPanelShown ? '' : ` ${videoPlayerStyle.mouseHidden}`}`}
       ref={containerRef}
       onMouseMove={handleShowPanel}
       onMouseLeave={() => setIsPanelShown(false)}
+      tabIndex={0}
+      onKeyDown={handleShortcut}
     >
       <video
         key={source}
@@ -85,14 +183,20 @@ export default function VideoPlayer({
       </figcaption>
       <div className={`${videoPlayerStyle.panel}${isPanelShown ? '' : ` ${videoPlayerStyle.hidden}`}`}>
         <VideoPlayerControlPanel
-          containerRef={containerRef}
           videoRef={videoRef}
           isPlaying={isPlaying}
           currentTime={currentTime}
           duration={duration}
-          setIsPlaying={setIsPlaying}
           setCurrentTime={setCurrentTime}
           bufferedProgress={bufferedProgress}
+          playVideo={playVideo}
+          pauseVideo={pauseVideo}
+          handleMuteVolume={handleMuteVolume}
+          handleFullscreen={handleFullscreen}
+          isMuted={isMuted}
+          setIsMuted={setIsMuted}
+          isFullscreen={isFullscreen}
+          handlePlayPause={handlePlayPause}
         />
       </div>
       {isBuffering
